Updated download_data.py
Renamed the method downloadMyAnimeData to myAnimeData. Added parameters such that the program calling it can store the data in the correct place.

Created the method hofAnimeData. This uses scrapes a given challenge page for eligible entries for that specific challenge. It then compiles the information in a dictionary, which it then returns.