Updated misc.py
Renamed method createHOFThemeDF to createHOFDF. Modified method so that it just creates a new merged DataFrame from the two passed in DFs.

Created method eligibleItems_Theme. Saves the filtered result of the given hofDataFrame based on the other given parameters in a new DataFrame and returns it.

Created method summaryOfChallenge. Returns a new DataFrame that describes the general status of a given challenge.