Updated misc.py.
Renamed loadMALJSONIntoDF to loadMALJSON_DF.

Rewrote the content in the loadMALJSON_DF method. Now this method returns a fully created DataFrame based on the column parameters and date instead of adding items to a DataFrame that was passed to the method. Added lines for an instance where the user would want to include the item's publishing period (start/end date).

Deleted extractDataColumn method.

Created combineMALDFs method.